chore(tauri/ui): rename components for clarity
